third_party/blink/web_tests/external/wpt/css/css-contain/content-visibility/content-visibility-svg-text-expected.txt

This is a testharness.js-based test.
[PASS] getComputedTextLength() should return nonzero values in a c-v:hidden subtree.
[FAIL] getCharNumAtPosition() should return nonzero values in a c-v:hidden subtree.
  Failed to execute 'getCharNumAtPosition' on 'SVGTextContentElement': parameter 1 is not of type 'SVGPoint'.
[PASS] getNumberOfChars() should return nonzero values in a c-v:hidden subtree.
[FAIL] getCTM() should return nonzero values in a c-v:hidden subtree.
  assert_equals: a expected 2 but got 1
[FAIL] getScreenCTM() should return nonzero values in a c-v:hidden subtree.
  assert_equals: a expected 2 but got 1
Harness: the test ran to completion.
